src/desktop_agent/pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nuvex-desktop-agent"
version = "0.1.2"
description = "NUVEX Desktop Agent — Windows automation client"
requires-python = ">=3.11"
dependencies = [
    "pywinauto>=0.6.8",
    "pywin32>=306",
    "mss>=9.0.1",
    "pynput>=1.7.7",
    "pyautogui>=0.9.54",
    "pystray>=0.19.5",
    "winotify>=1.1.0",
    "pyperclip>=1.8.2",
    "websockets>=12.0",
    "pillow>=10.3.0",
    "pydantic>=2.7.0",
]
